feat: 웹뷰 - 상세 진입 메시지핸들러 추가
미적용되어있던 디자인 컬러 적용
